server/requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==2.6.1
supabase
python-dotenv
sqlalchemy==2.0.15
psycopg2-binary==2.9.6
uuid

